+++ book-network/src/main/java/com/yladraoui/book/book/BookSpecification.java
@@ -4,7 +4,7 @@ package com.yladraoui.book.book;
 import org.springframework.data.jpa.domain.Specification;
 
 public class BookSpecification {
-    public static Specification<Book> withOwnerId(Integer ownerId){
+    public static Specification<Book> withOwnerId(Long ownerId){
         return ((root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("owner").get("id"),ownerId));
     }
 }
