+++ book-network/src/main/java/com/yladraoui/book/common/BaseEntity.java
@@ -25,7 +25,7 @@ import java.time.LocalDateTime;
 public class BaseEntity {
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     @CreatedDate
     @Column(nullable = false, updatable = false)
@@ -37,7 +37,7 @@ public class BaseEntity {
 
     @CreatedBy
     @Column(nullable = false, updatable = false)
-    private Integer createdBy;
+    private Long createdBy;
 
     @LastModifiedBy
     @Column(insertable = false)
